refactor: Moved abstractions into interfaces folders in event-sytem and core modules.
