+++ src/app/core/interfaces/ad-lib-piece-service.interface.ts
