+++ src/app/core/interfaces/entity-parser.interface.ts
