--- src/app/core/interfaces/rundown-event-service-interface.ts
@@ -1,3 +1,0 @@
-export interface RundownEventServiceInterface {
-  listenForRundownEvents(rundownId: string, onEvent: (event: any) => void): WebSocket
-}
