+++ src/app/core/services/basic-rundown-state.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, OnDestroy } from '@angular/core'
 import { BehaviorSubject, lastValueFrom, SubscriptionLike } from 'rxjs'
 import { RundownEventObserver } from './events/rundown-event-observer.service'
-import { Unsubscribe } from '../../event-system/services/event-observer.interface'
+import { Unsubscribe } from '../../event-system/interfaces/event-observer.interface'
 import { ConnectionStatusObserver } from '../../event-system/services/connection-status-observer.service'
 import { BasicRundown } from '../models/basic-rundown'
 import { BasicRundownService } from '../interfaces/basic-rundown-service.interface'
