+++ src/app/core/services/events/rundown-event-observer.service.ts
@@ -1,6 +1,6 @@
-import { EventConsumer, EventObserver, TypedEvent, Unsubscribe } from '../../../event-system/services/event-observer.interface'
+import { EventConsumer, EventObserver, TypedEvent, Unsubscribe } from '../../../event-system/interfaces/event-observer.interface'
 import { Injectable } from '@angular/core'
-import { RundownEventParser } from './rundown-event-parser.interface'
+import { RundownEventParser } from '../../interfaces/rundown-event-parser.interface'
 import { RundownEventType } from '../../models/rundown-event-type'
 import {
     RundownActivatedEvent, RundownAdLibPieceInserted,
