+++ src/app/core/services/http-rundown.service.spec.ts
@@ -2,7 +2,7 @@ import { HttpRundownService } from './http-rundown.service';
 import { HttpErrorService } from './http-error.service'
 import { instance, mock } from '@typestrong/ts-mockito'
 import { HttpClient } from '@angular/common/http'
-import { EntityParser } from './entity-parser.interface'
+import { EntityParser } from '../interfaces/entity-parser.interface'
 
 describe('HttpRundownService', () => {
   it('should be created', () => {
