+++ src/app/event-system/event-system.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RobustWebSocketFactory } from './services/robust-websocket.factory'
-import { EventObserver } from './services/event-observer.interface'
+import { RobustWebSocketFactory } from './factories/robust-websocket.factory'
+import { EventObserver } from './interfaces/event-observer.interface'
 import { WebSocketEventObserver } from './services/websocket-event-observer.service'
 import { ConnectionStatusObserver } from './services/connection-status-observer.service'
