+++ src/app/event-system/factories/robust-websocket.factory.ts
@@ -1,5 +1,5 @@
-import { RobustWebSocket } from './robust-websocket.service'
-import { ExponentiallyDelayedReconnectStrategy } from './exponentially-delayed-reconnect-strategy.service'
+import { RobustWebSocket } from '../services/robust-websocket.service'
+import { ExponentiallyDelayedReconnectStrategy } from '../services/exponentially-delayed-reconnect-strategy.service'
 import { Injectable } from '@angular/core'
 
 @Injectable()
