+++ src/app/event-system/interfaces/event-observer.interface.ts
