+++ src/app/event-system/interfaces/reconnect-strategy.interface.ts
