+++ src/app/event-system/services/websocket-event-observer.service.ts
@@ -1,7 +1,7 @@
-import { EventConsumer, EventObserver, TypedEvent, Unsubscribe } from './event-observer.interface'
+import { EventConsumer, EventObserver, TypedEvent, Unsubscribe } from '../interfaces/event-observer.interface'
 import { RobustWebSocket } from './robust-websocket.service'
 import { Injectable } from '@angular/core'
-import { RobustWebSocketFactory } from './robust-websocket.factory'
+import { RobustWebSocketFactory } from '../factories/robust-websocket.factory'
 
 @Injectable()
 export class WebSocketEventObserver implements EventObserver {
