+++ src/app/rundown/components/ad-lib-piece-identifier/ad-lib-piece-identifier.component.ts
@@ -1,6 +1,6 @@
 import {Component, Input} from '@angular/core'
 import {Identifier} from '../../../core/models/identifier'
-import {AdLibPieceService} from '../../../core/services/ad-lib-piece-service.interface'
+import {AdLibPieceService} from '../../../core/interfaces/ad-lib-piece-service.interface'
 
 @Component({
   selector: 'sofie-ad-lib-piece-identifier',
